Add files via upload
Stopwatch is a stylish, interactive web application built using HTML, CSS, and JavaScript. It features start, pause, reset, and lap functionalities with a glowing neon aesthetic and digital-style font, giving it a sleek retro-futuristic vibe. Designed for precision and visual appeal, this stopwatch is perfect for timing events with flair.